Move FusekiTestServer
